+++ src/plugins/google/api/drive/files/index.ts
@@ -1,1 +1,2 @@
 export { default as api } from './API';
+export { default as methods } from './methods';