+++ src/plugins/google/types.d.ts
@@ -1,0 +1,2 @@
+export type * from './api/types';
+export type * from './helpers/types';