+++ src/plugins/server/generators/database-handler/functions/index.ts
@@ -1,1 +1,3 @@
-export * from './edit-database';
+export * from './add';
+export * from './edit';
+export * from './delete';